Moved http constants to their own header file
