+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 08 08:12 PST
+Start testing: Nov 08 09:54 PST
 ----------------------------------------------------------
-End testing: Nov 08 08:12 PST
+End testing: Nov 08 09:54 PST
